product report added, controller modified

+++ shop/shop-repo/src/main/java/pl/rafalpieniazek/shop/repo/product/impl/ProductReportDTO.java
@@ -12,6 +12,6 @@ public class ProductReportDTO {
 
     private String category;
 
-    private int occurences;
+    private Long occurences;
 
 }
